feat: ajueste de layout para modulo de clientes
