+++ src/app/dashboard/layout.tsx
@@ -6,7 +6,7 @@ export default function DashboardLayout({ children }: LayoutProps) {
 	return (
 		<div className="min-h-screen">
 			Dashboard Layout
-			<section className="p-4">{children}</section>
+			<section className="p-2 md:p-4">{children}</section>
 		</div>
 	);
 }
